requirements.txt
flask>=2.3.0
nbformat>=5.9.0
python-docx>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.0
